ci: move the propose transaction task to a folder and add the example file
