+++ packages/stg-evm-v2/devtools/tasks/index.ts
@@ -138,7 +138,7 @@ import {
     TASK_STG_WIRE_TOKEN_MESSAGING_INITIALIZE_STORAGE,
     TASK_STG_WIRE_TREASURER,
 } from './constants'
-import './proposeTransactions'
+import './proposeTransactions/proposeTransactions'
 import './treasuryFee/proposeWithdrawTreasuryFee'
 import { checkResult } from './utils'
